the principle of least previlege

+++ app/Http/Controllers/Api/V1/TicketController.php
@@ -30,18 +30,11 @@ class TicketController extends ApiController
      */
     public function store(StoreTicketRequest $request)
     {
-        try {
-            // $user = User::findOrFail($request->input('data.relationships.author.data.id'));
-
-            // policy
-            $this->isAble('store', targetModel: Ticket::class);
-
-            // TODO: create ticket
+        // policy
+        if ($this->isAble('store', targetModel: Ticket::class)) {
             return new TicketResource(Ticket::create($request->mappedAttributes()));
-
-        } catch (AuthorizationException $authorizationException) {
-            return $this->error('You are not authorized to update this resource', 401);
         }
+        return $this->error('You are not authorized to update this resource', 401);
     }
 
     /**
@@ -71,16 +64,15 @@ class TicketController extends ApiController
         try {
             $ticket = Ticket::findOrFail($ticket_id);
 
-            // policy
-            $this->isAble('update', $ticket);
+            if ($this->isAble('update', $ticket)) {
+                $ticket->update($request->mappedAttributes());
 
-            $ticket->update($request->mappedAttributes());
+                return new TicketResource($ticket);
+            }
 
-            return new TicketResource($ticket);
+            return $this->error('You are not authorized to update this resource', 401);
         } catch (ModelNotFoundException $modelNotFoundException) {
             return $this->error('Ticket not found', 404);
-        } catch (AuthorizationException $authorizationException) {
-            return $this->error('You are not authorized to update this resource', 401);
         }
     }
 
@@ -90,15 +82,14 @@ class TicketController extends ApiController
         try {
             $ticket = Ticket::findOrFail($ticket_id);
 
-            $this->isAble('replace', $ticket);
+            if ($this->isAble('replace', $ticket)) {
+                $ticket->update($request->mappedAttributes());
+                return new TicketResource($ticket);
+            }
 
-            $ticket->update($request->mappedAttributes());
-
-            return new TicketResource($ticket);
+            return $this->error('You are not authorized to replace this resource', 401);
         } catch (ModelNotFoundException $exception) {
             return $this->error('Ticket not found', 404);
-        } catch (AuthorizationException $authorizationException) {
-            return $this->error('You are not authorized to replace this resource', 401);
         }
     }
 
@@ -111,11 +102,12 @@ class TicketController extends ApiController
             $ticket = Ticket::findOrFail($ticket_id);
 
             // policy
-            $this->isAble('delete', $ticket);
+            if ($this->isAble('delete', $ticket)) {
+                $ticket->delete();
+                return $this->ok('Ticket successfully deleted');
+            }
 
-            $ticket->delete();
-
-            return $this->ok('Ticket successfully deleted');
+            return $this->error('You are not authorized to replace this resource', 401);
         } catch (ModelNotFoundException $exception) {
             return $this->error('Ticket not found', 404);
         }
